设置地图功能增加

+++ src/views/room/Classify.vue
@@ -30,12 +30,20 @@
         </el-table-column>
       </el-table>
 
-      <el-dialog title="添加分类" :visible.sync="dialogVisible" width="38%" :before-close="handleClose">
+      <el-dialog title="添加分类" :visible.sync="dialogVisible" width="47%" :before-close="handleClose">
         <el-form :model="addClassificationForm" ref="addClassificationFormRef" label-width="100px">
-          <el-form-item :label="'分类名称'">
-            <el-input v-model="addClassificationForm.name" placeholder="请输入要添加的分类名称" style="width:200px"> </el-input>
+          <el-form-item
+            :label="'分类名称'"
+            prop="name"
+            :rules="{
+              required: true,
+              message: '分类名称不能为空',
+              trigger: 'blur',
+            }"
+          >
+            <el-input v-model="addClassificationForm.name" placeholder="请输入要添加的分类名称" style="width: 200px"> </el-input>
           </el-form-item>
-          <el-form-item v-for="(bed_type, index) in addClassificationForm.bed_type" :label="'床型'" :key="bed_type.index" :prop="'bed_type' + index">
+          <el-form-item v-for="(bed_type, index) in addClassificationForm.bed_type" :label="'床型' + (index + 1)" :key="bed_type.index" :prop="'bed_type' + index">
             <el-select v-model="addClassificationForm.bed_type[index].bed_label" placeholder="请选择" style="width: 120px">
               <el-option v-for="item in bed_label" :key="item.id" :label="item.bed_label" :value="item.bed_label"> </el-option>
             </el-select>
@@ -45,7 +53,7 @@
             <span> m</span>
             <el-input style="width: 60px" v-model="addClassificationForm.bed_type[index].num" type="number" :min="1"></el-input>
             <span> 张</span>
-            <el-button icon="el-icon-plus" circle  @click="addBed_type" size="mini"></el-button>
+            <el-button icon="el-icon-plus" circle @click="addBed_type" size="mini"></el-button>
             <el-button icon="el-icon-minus" size="mini" circle v-if="addClassificationForm.bed_type.length > 1" @click="removeBed_type(addClassificationForm.bed_type[index])"></el-button>
           </el-form-item>
         </el-form>
@@ -72,9 +80,6 @@ export default {
       },
       room_classify: [],
       bed_label: [],
-      addRules: {
-        long: [{ required: true, trigger: 'blur', message: '请输入！' }],
-      },
     }
   },
   created() {
@@ -103,14 +108,27 @@ export default {
         bed_type: [{ bed_label: '', long: '', wide: '', num: 1 }],
       }
     },
-    async addClassify() {
-      await createRoomClassify(this.addClassificationForm).then((res) => {
-        this.dialogVisible = false
-        this.GetClassifyList()
-        this.$message.success('添加分类成功!')
+    addClassify() {
+      this.$refs.addClassificationFormRef.validate(async (valid) => {
+        if (!valid) {
+          return
+        }
+        await createRoomClassify(this.addClassificationForm).then((res) => {
+          this.dialogVisible = false
+          this.GetClassifyList()
+          this.$message.success('添加分类成功!')
+        })
       })
     },
     async deleteClassify(id) {
+      const confiemResult = await this.$confirm('此操作将永久删除该分类, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning',
+      }).catch((err) => err)
+      if (confiemResult !== 'confirm') {
+        return this.$message.info('已取消删除！')
+      }
       await deleteRoomClassify({ id: id }).then((res) => {
         this.$message.success('删除分类成功!')
         this.GetClassifyList()
@@ -158,7 +176,7 @@ export default {
     .el-button {
       margin-left: 12px;
     }
-    .el-select{
+    .el-select {
       margin-left: 10px;
     }
   }
